refactor: create a monitoring module
