+++ src/mission_control/src/monitoring/cycles/register.rs
@@ -1,7 +1,7 @@
-use crate::cycles_monitoring::funding::init_funding_manager;
-use crate::cycles_monitoring::funding::register_cycles_monitoring;
-use crate::cycles_monitoring::store::heap::set_mission_control_strategy;
 use crate::memory::RUNTIME_STATE;
+use crate::monitoring::cycles::funding::init_funding_manager;
+use crate::monitoring::cycles::funding::register_cycles_monitoring;
+use crate::monitoring::store::heap::set_mission_control_strategy;
 use crate::types::interface::SegmentsMonitoringStrategy;
 use crate::types::runtime::RuntimeState;
 use crate::types::state::CyclesMonitoringStrategy;
