+++ src/mission_control/src/monitoring/cycles/unregister.rs
@@ -1,6 +1,6 @@
-use crate::cycles_monitoring::funding::init_funding_manager;
-use crate::cycles_monitoring::store::heap::disable_mission_control_monitoring;
 use crate::memory::RUNTIME_STATE;
+use crate::monitoring::cycles::funding::init_funding_manager;
+use crate::monitoring::store::heap::disable_mission_control_monitoring;
 use crate::types::runtime::RuntimeState;
 use ic_cdk::id;
 use junobuild_shared::types::state::SegmentId;
