+++ src/mission_control/src/monitoring/store/stable.rs
